requirements.txt
django==4.2.13
django-auditlog==3.0.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.0
django-modeltranslation==0.17.7
django-mptt==0.14.0
djangorestframework==3.15.2
drf-yasg==1.21.7
pillow==10.4.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
requests==2.32.3
redis==5.0.8
celery==5.4.0
django-ckeditor==6.7.1
firebase-admin==6.5.0
